Try to use action dispatch cookie jar first, fall back to new rack request
Rails 3.2.3 puts a cookie jar in @env, which we can use directly.
It also removes what we need from @env prior to this point, so the
old approach won't work.

Anything other than rails 3.2.x (other rack apps?) which mutate @env in a
way that populated action_dispatch.cookies will have to deal with the
consequences.