Don't set cookie over HTTP if secure cookie is on
Consider:

1. User signs in over HTTPS, gets secure only cookie
2. User browses away.
3. User returns to site over HTTP.

Because this new connection is over HTTP and the cookie is set to secure
only, the client will not send it. Clearance will see that the cookie is
missing create a new cookie, with no value, causing the user to be
signed out.

This change short circuits the cookie setting if the cookie is to be
secure_only and the request is not over HTTPS.