Use available query string when redirecting
If the return_to url has a query string attached to it, we want to be
sure we are including it when clearance redirects.

URI::HTTP has the wonderful `request_uri` method that returns us the
path and query string together, but our session variable stores only
`fullpath` which URI parses as URI::Generic. That means we have to
assemble the path and query string ourselves.