Allow easy overriding of clearance controllers
This is particularly useful for overriding the controller used in the
sign_in, sign_up, and sign_out routes. Rails 4 no longer allows
overriding named routes (an exception is raised), so having `sign_in`
use your customized controller (even when it was named `users`) required
adding an unnamed (`as: nil`) route that matched the same path and went
to a different controller.

This difficulty is not a problem for resourceful routes. That is, adding
your own `resoures :passwords` to your routes file will override the
routes as expected. However, configuration was added for all controllers
for the sake of consistency and because it allows the user to change the
controller without having to know which actions the resources need to
have available for Clearance.