User can request a password reset
This change introduces a `PasswordReset` model to store the token
generated for a user resetting a password. It also re-introduces
ClearanceMailer which sends the reset notice to the user.

The install generator was updated to copy a migration that creates the
`password_resets` table (though this name can be derived from an
optional argument to the generator).

Configuration options were introduced to drive the password reset
process:

* `mailer_sender` is the from address for the mailer.
* `password_reset_class` is the class to use for password resets.
* `password_reset_delivery_method` is a lambda used in the
  `PasswordResetsController` to deliver the mail and is intended to be
  overridden in order to introduce background mail delivery.