Don't raise `NoMethodError` in `sign_in_as` helper
I was overriding `sign_in` in a subclassed controller (ApplicationController was
public, but a subclass was used for all pages requiring authentication), and
experienced problems during testing where exceptions were being raised by
Clearance's testing helpers calling the protected method `#sign_in` on the
controller.

Changed the `sign_in_as` helper to use `Object#send`, which will not raise
no method exceptions if a controller's `#sign_in` method is protected, which as
a non-action controller method it should be.

Close #424.