Remove configuration update made by generator
This code doesn't work. It results in a mangled initializer. We can
address this in a future release if needed, but I  think it might be
best for it to be two separate steps anyway.

Also changed the indentation in the routes template. It looks off there
but it results in nicely formatted routes when inserted.