Don't set a blank remember token
There's no sense in setting a remember token cookie if the value of that
cookie is nil. In fact, this causes a problem for sites that bounce
users between HTTP and HTTPS connections for when the user is signed in
versus signed out. Prior to this change, if Clearance's `secure_cookie`
option was on the users remember token would be completely wiped out
when they hit an HTTP page. Even returning to HTTPS would not restore
the user's token. With this change in place, the HTTP response would
simply not set a remember token cookie, but switching back to HTTPS
would allow the client to send its previously existing cookie.

There was an existing test for this, but it was passing incorrectly. The
`set_cookie` matcher was not negating the way we wanted it to. On the
test, the header values were:

Fixes #338.