Fix test when should not set a empty cookie
The test was probably meant to make sure remember_token was not there,
however the matcher was not working on the negative case.

On the test the headers values was:
{"Set-Cookie"=>"remember_token=; path=/; expires=Wed, 06 Jan 2016
15:19:48 -0000"}
Which was not matching the `set_cookie('remember_token')` matcher,
because the expires_at was set, thus not matching the matcher, and
hiding that remember_token was there.