Do not validate user when resetting password
User records can be invalid for many reasons that are not central to
Clearance's concerns. An application may decide at some point to begin
requiring all user records to have a zip code. They reasonably assume
this will effect only new registrations or existing users who attempt to
edit their profile. Unfortunately, it would have also prevented users
from resetting their password because we were running validations as
part of that operation. With zip code not present on the password reset
form, the user would have no way to complete the password reset.

This change still calls `valid?` but considers only errors that are on
the `password` field when deciding whether to proceed. This allows for
additional validations on `password`, such as length. Ultimately, the
save is done without validations assuming `password` itself is valid.

Closes #406