Enable forgery protection on sessions#create
This line has existed since 2008, and yet I can determine no
justification for it. It seems to me that we *would* want CSRF
protection on `session#create`.

On its own, skipping CSRF protection in just this single action doesn't
seem particularly useful to an attacker. Additional vectors (such as an
overly-permissive CORS header) would have to be present to make use of
this, but at that point far more interesting attacks would be possible
on any cookie-based auth system.