Document Clearance user module
Several areas of clearance, the user module being one, contain methods
that are marked as private in ruby but are actually part of the public
API for the gem. For instance the private methods in the User module are
private so that they aren't exposed publicly in the host application's
user model, but they are certainly available for use to developers.

For this reason, this change include changes to the yard options to
allow documentation of private methods. If a method in Clearance is
truly internal API then it should be marked with `@api private`. The
yard options will skip these.