Clean up indentation of generated routes
The `route` generator action provided by rails assumes a single line
route injection. The first line is automatically indented, but the
following lines are not.

I thought about injecting the routes one line at a time, but that
results in weird output from the generator because we format some long
lines in the routes to avoid going over 80 characters.

I decided to override the `route` action provided by default rails
generators to do exactly what we need. This also has the benefit of
cleaning up the output of the generator as well.