Prevent `skip_before_filter` errors in Rails 5
Rails 5 will error if `skip_before_filter` is called with a filter name
that has not yet been added. We can't always control if the `authorize`
filter is in the inheritance chain, but we always want to be sure if it
is, we skip it.

Adding `raise: false` allows us to opt out of the new behavior without
impacting older versions of rails.