Eliminate HTTP method deprecations from tests
Rails 5 deprecates calling HTTP action methods with positional arguments
in favor of keyword arguments. However, the keyword argument form is
only supported in Rails 5+. Since we support back to 3.1, we need some
sort of shim to avoid super noisy deprecations when running tests.

I tried to shim this the other way - update tests to the new format, and
shim older versions of rails, but that causes failures in older versions
of Rails. Meanwhile, shimming in this direction actually fixed failures
I was seeing on Rails 5. It appears the Rails 5 code for rewriting
positional arguments to keyword arguments is broken. See: https://github.com/rails/rails/issues/23643