Always use inline adapter in feature specs
Rails 5 defaults the ActiveJob adapter to `async` which makes it
difficult to test the password reset feature in a manner that supports
Rails 4.2 and 5.0 (and older versions as well, of course).

The simplest fix, for now, is to use the inline adapter in our tests so
we can observe the side effects we care about.