Create DeferredSignInUser class
This class wraps a user object and a password string, thus allowing sign
in guards to have an unauthenticated user object to work with. It's
meant to be part of the defer_sign_in_password_check feature, which
passes unauthenticated AND existing user objects to sign in guards for
auditing purposes.

Clearance's default behavior is to represent an unauthenticated user
(which password is wrong) with nil, so sign in guards end up receiving a
nil user when password is incorrect.

This new feature should be useful to allow clients to implement other
features, like Account Lock Guards, which require working with a user
object even if password has not been checked yet.

This class overrides the present? method, which is currently used by
clearance to evaluate whether a user is signed in or not. This present?
method only returns true if a user is present AND signed in, thus it
keeps up with Clearance's default behavior.