Correctly track dirty state of encrypted password
Password validations only run when `skip_password_validation?` returns
false. The idea of that method is that it returns true whenever the user
requires a password and a change to their password was attempted.
Previously, we relied on the developer explicitly having set
`password_changing = true` on the instance to detect an attempt to
change the password.

This worked fine when the developer used `update_password` to change an
existing users password. However, using other methods to update a
password or even creating a new user (like in the `users#create` action)
did not set this attribute.

This was okay for our built-in validations (just presence) because of
the way `password=` was implemented. However, users that added their own
validations (like validating password length) and properly added the
`unless: :skip_password_validation?` modifier to their validation found
that their validations were not running for new users.

To fix this, we rely on ActiveRecord/ActiveModel's built-in dirty
tracking to check if encrypted password has changed rather than using a
manual attribute. We further tell the user model that calling
`password=` at all will automatically mark the encrypted_password as
dirty, regardless of what the password strategy ultimately does.

Making the change in `User` and calling `super` means that all password
strategies -- even custom strategies -- get this improved behavior.

Now that we're using dirty tracking properly, we can deprecate the
`password_changing` attribute, as it is no longer used.

Closes #664