Reset Rails session on sign out
Signing out currently removes only the dedicated Clearance remember
token cookie. The Rails session itself is left in tact because, as far
as Clearance is concerned, there is no user-specific information stored
there. This is true, but is also naive. It's common for Rails developers
to store user-specific information in `session`.

Resetting the session will also cause the CSRF token to be rotated. CSRF
issues typically involve signed-in users and Clearance sign out does not
leave the user signed in, so there's not much value to a CSRF attack
against a session in this state. Still, rotating the token is a best
practice.