Reset CSRF token on sign in
By rotating this token, we can prevent session fixation attacks. This is
the simplest explanation of session fixation that I've found, and I do
not want to recreate it here: https://www.owasp.org/index.php/Session_fixation

By rotating the CSRF token on sign in, the attacker will be unable to
make authenticated requests using the known session ID as they will not
have access to the new CSRF token without an additional security
vulnerability.

While I believe this should cause most Clearance users no harm if
enabled by default, there's a chance that it may, particularly if sign
in is done via AJAX, as additional form submissions after sign in will
not have the updated CSRF token.

For that reason, this feature is behind a configuration setting. The
configuration is `nil` by default, which will generate a warning that
encourages the user to set it to `true`, but the warning can be silenced
by setting the configuration to any value.