Remove unused nested `create` route
In the routes file we have two `create` routes defined for the
`password` resource. One on the `password` resources and another
one as nested `member` routes for users, but we only use the
`resources` one to reset user's password.

This commit removes the unused nested `create` routes for the
password resource.