Return authentication cookie only if authentication is performed
It's not safe to unconditionally write the authentication cookie on
every response. If the application chooses to use http cache on a given
endpoint, and the first hit to that endpoint is performed by an
authenticated user, its cookie will be cached and reused as a caned
response for any other incoming request.

The authentication cookie should be set only if the application has used
authentication during this request.