Fix password related issue for old/new token
Recently, to prevent password reset token via HTTP, there was
an update to store the reset token in session and then redirect
to the same edit page without token in the params, this works
great but the way we find user introduced a new issue like 749

If the user request a password reset and click on that link and
for some reason that reset was unsuccessful and then the user
request another link then it does not work and it keep sending
the user to the `password/new` page again and again.

Why is that happening?

The password `#edit`, and `#update` actions has before filter
to ensure the token is valid for an existing user, and in the
underneath the `ensure_existing_user` is trying to find a user
by the session token / parameter token.

But it gives precedence to the session token first and in above
scenario, although the user has a new token in the URL but the
session still holds the old token which is not valid anymore
and that creates the indefinite loop.

This commit adds a little tweaks by switching the order and
when a URL contains the `token` parameter then it updates the
session reset token and then redirect to the same edit page.
This way it updates the session &everything else stays same

Fixes: #749