Allow fragment in `session[:return_to]` value
Under ordinary operation, clearance automatically sets the
`session[:return_to]` value when a non-signed-in user hits a protected
page. By spec, this value won't have a fragment because the fragment is
never sent to the server (it's all handled client side).

However, it's possible that developers will want to manually set a
`return_to` value that has a fragment. Prior to this change, this would
always be stripped out. This change fixes that by preserving the
fragment when it is present and eliminating any dangling `#`s.