Fix broken `deny_access` rspec matcher
In a previous commit deprecating old rails behavior, we mistakenly deprecated
the newer behavior instead of the older behavior.

This change fixes our deprecation mistake, and adds some basic flexing of this
matcher in the existing spec suite to ensure that we're at least running it with
a flash check involved.