Update email_validator gem and relax validations
The gem we use for email format validation recently changed its approach to
validation and radically relaxed the gem validation.

This blog post was part of the rationale: https://medium.com/hackernoon/the-100-correct-way-to-validate-email-addresses-7c4818f24643

This change updates the version of the gem we are using, and as a side effect,
relaxes the requirements for valid email addresses by default in clearance user
models.