Clean up sql output in generated migration
Adding the `squish` here has the effect of slightly cleaning up the eventual SQL
generated by the migration we are helping create here. Before this change the
migration would be full of newlines (or `\n`). Post-change, each sql statement
should come out cleaner in the logs with the line breaks and tabs/spaces
replaced by just one space each.

The SQL that is run stays the same, it just looks slightly nicer in the log
output.