Add parameter stripping to BackDoor
Before, we were leaving the "as" parameter in the query string when
accessing the back door. Doing so was causing problems downstream when
the consuming application was using [Strong Parameters]. If the app had
not listed the "as" parameter as a permitted parameter, then tests would
fail. We added functionality to the back door to strip out the "as"
parameter before calling the next middleware.

[Strong Parameters]: https://api.rubyonrails.org/classes/ActionController/StrongParameters.html