Delete remember cookie when domain is a Proc
The `cookie_domain` setting [can be set to a callable][1] to allow
support for multiple domains.

When we generated the [options used to delete the cookie][2] we passed
the setting value directly. When the setting was a callable this passed
the callable rather than the value.

This commit resolves the domain and uses the value in the options.

[1]: https://github.com/thoughtbot/clearance/blob/c3da4f62a7300b177d3b7d02b246144f6d473372/README.md#multiple-domain-support
[2]: https://api.rubyonrails.org/classes/ActionDispatch/Cookies.html