Add sqlite3 to Gemfile and remove from Appraisals (#906)
Closes #884.

Before this commit, running the `bin/setup` script was breaking. This was
because `bundle exec rake dummy:db:reset` command used the top level Gemfile
where the sqlite3 gem was not specified. It was only specified in the Appraisal
Gemfile.

This commit specifies the sqlite3 gem in the top level Gemfile as well.

Co-authored-by: Daniel Colson <danieljamescolson@gmail.com>
Co-authored-by: Eebs Kobeissi <ebrahim.kobeissi@gmail.com>