Separate out appraisal versions
Before this commit we were generation appraisals in a loop. This may
have worked fine when every version of Rails worked with the same set of
gems, but now that we need to conditionally constrain gem versions based
on the Rails version this no longer makes sense.

This commit declares each appraisal version separately, specifying only
the specific gem versions that are different from the top-level Gemfile
for that specific appraisal.

As an added bonus, moving some of these gems into the top-level Gemfile
made it possible to run the tests without appraisal as well.