Bump all dev dependencies
This commit remove the version constraints for our development
dependencies and bumps them all in both the top-level Gemfile and the
appraisal Gemfiles (via `appraisal update`).

This has the nice effect of getting rid of lots of Ruby 2.7 keyword
argument warnings in our tests. The Ruby 2.7 + Rails 6 tests is now
nothing but green dots.