Reduce timing variance when running benchmarks
The test that benchmarks how long it takes to authenticate when a user
exists vs when a user does not exist would occasionally fail on CI. We
believe this to be caused by minor timing variance.

This commit updates the spec to run the authentication 10 times each
rather than a single time. This will help to even out the observed
variance.