Revert case sensitivity for email uniqueness
In #889, `case_sensitive` setting of the `uniqueness` validator was set
to `false` to prevent deprecation warnings from Rails.

However, the setting should have been set to `true` to preserve
backwards-compatibility.

The deprecation warning said:

> To **continue** case sensitive comparison
> on the :email attribute in User model, pass case_sensitive: true
> option explicitly to the uniqueness validator.

Setting it to `false`, made ActiveRecord generate User queries on
`LOWER(email)` instead of on `email`. The apps that didn't have an index
on `LOWER(email)` would have faced performance issues.

This was mentioned in an old commit that originally made the switch from
`false` to the implicit `true`:
https://github.com/thoughtbot/clearance/commit/56f0f0489a138d64794136e133c5848183c576d5