Increase time range for authentication spec
In commit ce17d2d6d861bb3415602c388bbf5bda8b5fceec some specs were added
to verify that the time it takes to authenticate users that already
exist and those that don't is within 0.001s. This causes the tests to
fail randomly given that the time range is too precise. We are
increasing this range to 0.01 to avoid flaky tests.