Explain how to use the testing backdoor outside of Rails routes
Some projects may want to use the backdoor, but may not be using Rails routes in their system tests. One example is an app that hosts a React frontend, but still uses Clearance cookies for auth. A system test with Capybara may visit a React route like this: `visit "/payment_method"`. 

The existing documentation didn't show how to use the backdoor with this kind of route.