Call dynamic README workflow (#1004)
We want to have a way to edit our README footer at one place and have
the changes from there be propagated to our repos.

By adding this snippet in the README, we call this reusable workflow:
https://github.com/thoughtbot/templates/blob/main/.github/workflows/dynamic-readme.yaml
that renders and updates the README footer dynamically.

This workflow has been updated to accommodate repositories that have
main branch rules. To generate the README dynamically, you’ll need to
merge a PR with the README changes. That will trigger the GitHub workflow.
The PR won’t be generated if the footer does not have any differences.
It will only be open if there’s an update to the footer snippet.